FIX: pin as string

See merge request humansis/mobile-apps/vendor-app!137

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/main/checkout/fragment/CheckoutFragment.kt
@@ -427,7 +427,7 @@ class CheckoutFragment : Fragment(), CheckoutFragmentCallback {
                     pin.length == 4 -> {
                         dialog?.dismiss()
                         findNavController().navigate(
-                            CheckoutFragmentDirections.actionCheckoutFragmentToScanCardFragment(pin.toInt())
+                            CheckoutFragmentDirections.actionCheckoutFragmentToScanCardFragment(pin)
                         )
                     }
                     pin.isEmpty() -> {
